third_party/blink/web_tests/css3/calc/font-size-expected.txt

Tests that CSS3 calc() can be used with the font-size property

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(document.getElementById("calc-percent"), null).fontSize is "40px"
PASS getComputedStyle(document.getElementById("calc-percent-pixels"), null).fontSize is "40px"
PASS successfullyParsed is true

TEST COMPLETE
